requirements.txt
numpy==1.26.4
discord.py==2.6.4
yt-dlp==2025.12.8
yt-dlp-ejs==0.3.2
PyNaCl==1.6.1
discord-py-interactions==5.15.0
discord-components==0.0.0.1